services/api/pyproject.toml
[project]
name = "botcheck-api"
version = "0.1.0"
description = "BotCheck Control Plane — FastAPI service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "pydantic-settings>=2.4",
    "python-jose[cryptography]>=3.3",
    "passlib[argon2]>=1.7",
    "arq>=0.26",
    "httpx>=0.27",
    "livekit-api>=0.7",
    "openai>=1.0",
    "anthropic>=0.40",
    "aioboto3>=13.0",
    "prometheus-client>=0.20",
    "croniter>=3.0",
    "qrcode>=7.4",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "opentelemetry-instrumentation-httpx>=0.40b0",
    "opentelemetry-instrumentation-openai-v2>=2.0b0",
    "structlog>=24.1",
    "pyyaml>=6.0",
    "botcheck-http-client",
    "botcheck-observability",
    "botcheck-scenarios",
]

[tool.uv.sources]
botcheck-http-client = { workspace = true }
botcheck-observability = { workspace = true }
botcheck-scenarios = { workspace = true }

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "aiosqlite>=0.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["botcheck_api"]
